Don't check if the line is an import and use a no wrap search
As we start the search when the first class definition, there should not
be anymore imports. We make sure to have a non wrap search in order to
not have to scan the list of imports.

The negative lookbehind for non import lines was quite inefficient,
removing it should bring a boost in performance.