Ignore wildcard imports
What
===
Ignore wildcard imports.

Why
===
The plugin at the moment doesn't distinguish between wildcard and exact
imports and when the plugin doesn't find the use of an asterisk in the
file it assumes the wildcard import should be removed. It is difficult
to determine if wildcard imports are unused so it's better if we just
ignore them. A lot of Java developers do not use wildcard imports, but
if you happen to jump into a file that has them and this plugin is
enabled it removes them, which is inconvenient. Chances are we just want
to leave it and not change the import.

Also
===
The lack of a line ending marker in the regex created issues with the
asterisk in tests I ran, so I also added the $ at the end and the \s
whitespace characters around the ; to ensure stray whitespace on the end
of import lines doesn't prevent the plugin from working.